[ME] Separate property builders, dropdown styling and virtual-list helpers
