+++ src/MaterialEditor.Base/MaterialEditor.Base.projitems
@@ -9,12 +9,28 @@
     <Import_RootNamespace>MaterialEditor.API</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)Assets\MaterialTextureExport.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Manifest\ShaderDefinitionLoader.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\Virtualization\VirtualListCachePolicy.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\Virtualization\VirtualListViewPool.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\Sections\Properties\PropertyRowBinding.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\Styling\MaterialEditorDropdownPopupLayout.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PluginBase.Configuration.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PluginBase.ShaderModels.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PluginBase.ShaderPropertyParsing.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)UI\Sections\Properties\PropertyRowBuilders.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Manifest\ShaderDefinitionLoader.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Assets\MaterialTextureExport.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\Sections\Properties\ExtensionPropertyRowBuilder.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\Sections\Properties\PropertyTextureRowBuilder.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\Sections\Properties\PropertyValueRowBuilder.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\Styling\MaterialEditorDropdownCaptionFitter.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\Styling\MaterialEditorDropdownPopupStyle.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\Styling\MaterialEditorDropdownItemStyle.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\Styling\MaterialEditorTextStyleState.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\Styling\MaterialEditorPanelStyleState.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\Styling\MaterialEditorControlStyleState.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\Styling\MaterialEditorScrollStyleState.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\Styling\MaterialEditorGraphicStyleState.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\Styling\MaterialEditorOutlineStyleState.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\Styling\MaterialEditorThemeRepaintCoordinator.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Operations\MaterialTextureImportTransaction.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Operations\MaterialTextureSnapshot.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Operations\MaterialTextureImportCommit.cs" />
